requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.0
PyNaCl==1.5.0
ffmpeg-python==0.2.0
requests==2.31.0
aiohttp==3.9.1
